Added note-switching feature
Changelog:
Certain config files are now to be located inside base/configs folder. These are sounds.ini, rpccharlist.ini, and filesabstract.ini (used to determine which files/notes the user has loaded)
Added note-switching widget. It consists of a set_notes button, which pops it up. As well as a note_edit button to edit which note to select, note_select button to select which note to edit, and note_delete button to delete a note. The parent widget to which an image can be set is note_area. Also, note_select has _selected counterpart.
To locate the widget in courtroom_design.ini, refer to note_area and set_notes_button.
Some minor changes.