Don't fix what's doubly broken: Emote dropdown.
Made it so that emote dropdown only responds to activation, instead of
index changing, i.e., reverted to previous behaviour.

Emote changing already had a failsafe for the emote dropdown index
changing, which activated even when you clicked on the emotes. That made
it so that you'd get the Pre button ticked on emotes that didn't have
'em, and vice versa.