The charselect's "shadows" correctly update based on what chars are taken.
This was purely a graphical bug. The characters were correctly
recognised as taken by the client, but there was no way to update the
"taken-shadow" over their icons.
Which meant that they were locked into the way they were when the user
first joined the server.

As a result of this, a `CharsCheck` package from the server will
correctly display the taken characters to the client in the character
selection.