Moved realisation to `start_chat_ticking()`.
Since the realisation could have been called from two paths that were
previously one and the same (previously, the pre ending was the same as
the text starting -- with the inclusion of non-interrupting preanims,
these became different), in some select cases, the realisation could
have played twice.

As a solution to this, handling realisation is now done when the text
starts ticking, as that guarantees that the realisation happens only
once.