(unfinished) Continue full refactoring
At some point I was able to produce a working build, but I
unfortunately did not commit the changes to that point. It was
not that interesting anyway - half the UI was on the original
window, and the other half was on another window.

This is the culmination of some months of work. I still have more
unstaged changes.