Fix widget and crash bugs
The code is consistently compiling now, and it's possible to connect to
a server again!

The three largest issues right now are the fact that the character
selection dialog is not opening, there are too many widgets being
crammed into the main window at the same time, and the stability
of the game is back to zero. I'm sure these are relatively trivial
problems compared to the total amount of time that has been placed in
this entire project.