Fix crash on closing settings dialog
This is why I don't ever want to mess with a dialog that was not made
with the designer. The bug was very obscure and its crash error
("Detected QWeakPointer creation in a QObject being deleted")
was very far from the actual source of the issue. It would only crash
on the debug build and not on the release build, too, and the only
symptom on the release build would be that the entire Options dropdown
menu would be removed.