Fix connect button not enabling
Apparently this had never been staged, and while everyone was having
problems with the connect button, everything was working just fine for
me.

This enables the connect button a bit earlier in the handshake process
than in 2.6, since there are some older servers out there that don't
send FL (and so the connect button would never get enabled). This might
lead to race conditions, and sometimes the progress bar does hang,
but you can just cancel and reconnect - I haven't encountered a crash
so far.