Layout persistence
The courtroom layout can now be saved to a file. Even with no action,
the layout is autosaved to a file called last_layout.dat every 30
seconds as well as on exit. When there is no autosaved layout, the
theme's default layout will be loaded (though, frankly, I should
probably store layout autosaves at the theme level instead of globally.)