Fix some IC input issues; remove MS chat from courtroom
- The position dropdown and preanimation checkboxes had not been wired
correctly, so they weren't doing
 anything.
- Similarly, the MS chat pane was not doing anything, so I removed it
altogether.

Next on the chopping block is the mostly useless casing UI and netcode
that has given me such a headache refactoring (due to the number of
parameters, one for each role).

The courtroom's done for now - if time allows, I might give the lobby UI
 a whirl :^)