Add a new way of sending login information to the server.
This commit adds LoginDialog, a dialog box for inputting a username and password, and sending it to the server, removing the need to use sever commands (and thus, removing the potential to accidentally leak moderator passwords). The login dialog will send an AUTH packet to the server, containing the username and password. (If a server doesn't support usernames, the username field can be left empty).

* Rename debug_functions to dialog_functions
* Add LoginDialog
* Add a new button to courtroom.cpp for opening the login dialog.