Remove gender as a valid ini option
Given that it has now been around a year that blips has replaced gender as the primary option to set a character blip it has been more than enough time for servers to adjust their content to the new standard.

With in-client debug and warnings on the horizon this change is not as much of an issue as it was with prior clients as users can check error messages directly.