fix evi ordering while we're at it
Evidence is saved to .ini files as a QSettings with groups named after their index in the list. When it's loaded back in from the file by getting childGroups, however, it's sorted in that lovely machine way that puts 10-19 after number 1 (ONE) but before number 2 (TWO). This is an attempt to fix it by human-like sorting, and testing has produced satisfactory results.